requirements.txt
fastapi
uvicorn
httpx
gunicorn
fastapi-cache2
